requirements.txt
fastapi
uvicorn
pymupdf
pdfplumber
pdfminer.six